move dataframe plotting utility script to utils
