+++ asapdiscovery/docking/modeling.py
@@ -1,6 +1,6 @@
 from openeye import oechem, oedocking, oespruce
 
-from .datasets.utils import (
+from asapdiscovery.data.openeye import (
     load_openeye_pdb,
     load_openeye_sdf,
     split_openeye_mol,
